test: move tests to sit alongside what they're testing
