+++ src/diff.test.ts
@@ -1,12 +1,12 @@
 import { mocked } from 'ts-jest/utils';
 import path from 'path';
-import { getBaseBuild, matchConfigs } from '../src/diff';
-import { COMMIT, fakeProcess } from './fixtures';
-import { mergeBase, revList } from '../src/git';
-import getConfigs, { getBuildkiteInfo } from '../src/config';
+import { getBaseBuild, matchConfigs } from './diff';
+import { COMMIT, fakeProcess } from '../test/fixtures';
+import { mergeBase, revList } from './git';
+import getConfigs, { getBuildkiteInfo } from './config';
 
-jest.mock('../src/git');
-jest.mock('../src/buildkite/client');
+jest.mock('./git');
+jest.mock('./buildkite/client');
 
 const mockRevList = mocked(revList, true);
 mockRevList.mockImplementation(() => Promise.resolve([COMMIT]));
@@ -31,7 +31,7 @@ describe('matchConfigs', () => {
 
   it('matches changed files against configs', async () => {
     process.env = fakeProcess();
-    process.chdir(path.resolve(__dirname, 'projects/simple'));
+    process.chdir(path.resolve(__dirname, '../test/projects/simple'));
     const result = matchConfigs('foo', await getConfigs(), ['foo/abc.js', 'foo/README.md', 'bar/abc.ts', 'baz/abc.ts']);
 
     expect(result[0].changes).toStrictEqual([]);
